edit & delete you own posts
You can only edit & delete your own posts
You can't edit & delete outher peoples posts

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,15 @@ use App\post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -75,6 +84,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('posts')->with('error', 'Unauthorized page');
+        }
         return view('post.edit')->with('post', $post);
 
     }
@@ -106,6 +120,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('posts')->with('error', 'Unauthorized page');
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
